Ai_Models/requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
paho-mqtt>=1.6.1
requests>=2.31.0
google-genai
fastapi
uvicorn
pydantic

